Refactor backend user and group routes
