+++ .gitignore
@@ -140,3 +140,5 @@ dist
 # Vite logs files
 vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
+cookies.txt
+cookies.txt
