+++ packages/backend/models/group.js
@@ -2,14 +2,14 @@ import mongoose from "mongoose";
 
 const groupSchema = new mongoose.Schema(
   {
-    // User who created the group
+    //user who created the group
     owner: {
       type: mongoose.Schema.Types.ObjectId,
       ref: "User",
       required: [true, "owner is required"],
     },
 
-    // Users in the group
+    //users in the group
     users: {
       type: [
         {
@@ -20,14 +20,13 @@ const groupSchema = new mongoose.Schema(
       default: [],
     },
 
-    // Group goal from the written requirement
+    //group goal
     groupGoal: {
       type: Number,
       default: 0,
       min: [0, "groupGoal cannot be negative"],
     },
 
-    // Hours from the diagram
     hours: {
       type: Number,
       default: 0,
